Struct boa_interner::fixed_string::FixedString

pub(crate) struct FixedString<Char> {
    inner: Vec<Char>,
}

Fields
inner: Vec<Char>
Implementations

impl<Char> FixedString<Char>

pub(crate) fn new(capacity: usize) -> Self

Creates a new, pinned FixedString.

pub(crate) fn capacity(&self) -> usize

Gets the maximum capacity of the FixedString.

pub(crate) fn is_empty(&self) -> bool

Returns true if the FixedString has length zero, and false otherwise.

impl<Char> FixedString<Char>where
Char: Clone,

pub(crate) unsafe fn push(
&mut self,
string: &[Char]
) -> Option<InternedStr<Char>>

Tries to push string to the FixedString, and returns an InternedStr pointer to the stored string, or None if the capacity is not enough to store string.

Safety

The caller is responsible for ensuring self outlives the returned InternedStr.

pub(crate) unsafe fn push_unchecked(
&mut self,
string: &[Char]
) -> InternedStr<Char>

Pushes string to the FixedString, and returns an InternedStr pointer to the stored string, without checking if the total capacity is enough to store string, and without checking if the string is correctly aligned.

Safety

The caller is responsible for ensuring that self outlives the returned InternedStr and that it has enough capacity to store string without reallocating.