Struct boa_interner::raw::RawInterner

pub(crate) struct RawInterner<Char> {
    symbol_cache: FxHashMap<InternedStr<Char>, usize>,
    spans: Vec<InternedStr<Char>>,
    head: FixedString<Char>,
    full: Vec<FixedString<Char>>,
}

Raw string interner, generic by a char type.

Fields
symbol_cache: FxHashMap<InternedStr<Char>, usize>spans: Vec<InternedStr<Char>>head: FixedString<Char>full: Vec<FixedString<Char>>
Implementations

impl<Char> RawInterner<Char>

pub(crate) fn with_capacity(capacity: usize) -> Self

Creates a new RawInterner with the specified capacity.

pub(crate) fn len(&self) -> usize

Returns the number of strings interned by the interner.

pub(crate) fn is_empty(&self) -> bool

Returns true if the interner contains no interned strings.

impl<Char> RawInterner<Char>where
Char: Hash + Eq,

pub(crate) fn get(&self, string: &[Char]) -> Option<usize>

Returns the index position for the given string if any.

Can be used to query if a string has already been interned without interning.

pub(crate) fn intern_static(&mut self, string: &'static [Char]) -> usize

Interns the given 'static string.

Returns the index of string within the interner.

Note

This is more efficient than RawInterner::intern, since it avoids storing string inside the interner.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub(crate) fn index(&self, index: usize) -> Option<&[Char]>

Returns the string for the given index if any.

unsafe fn next_index(&mut self, string: InternedStr<Char>) -> usize

Inserts a new string pointer into spans and returns its index.

Safety

The caller must ensure string points to a valid memory inside head (or only valid in the case of statics) and that it won’t be invalidated by allocations and deallocations.

impl<Char> RawInterner<Char>where
Char: Hash + Eq + Clone,

pub(crate) fn intern(&mut self, string: &[Char]) -> usize

Interns the given string.

Returns the index of string within the interner.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.