Struct boa_interner::Sym

pub struct Sym {
    value: NonZeroUsize,
}

The string symbol type for Boa.

This symbol type is internally a NonZeroUsize, which makes it pointer-width in size and it’s optimized so that it can occupy 1 pointer width even in an Option type.

Fields
value: NonZeroUsize
Implementations

impl Sym

pub const EMPTY_STRING: Self = _

Symbol for the empty string ("").

pub const ARGUMENTS: Self = _

Symbol for the "arguments" string.

pub const AWAIT: Self = _

Symbol for the "await" string.

pub const YIELD: Self = _

Symbol for the "yield" string.

pub const EVAL: Self = _

Symbol for the "eval" string.

pub const DEFAULT: Self = _

Symbol for the "default" string.

pub const NULL: Self = _

Symbol for the "null" string.

pub const REGEXP: Self = _

Symbol for the "RegExp" string.

pub const GET: Self = _

Symbol for the "get" string.

pub const SET: Self = _

Symbol for the "set" string.

pub const MAIN: Self = _

Symbol for the "<main>" string.

pub const RAW: Self = _

Symbol for the "raw" string.

pub const STATIC: Self = _

Symbol for the "static" string.

pub const PROTOTYPE: Self = _

Symbol for the "prototype" string.

pub const CONSTRUCTOR: Self = _

Symbol for the "constructor" string.

pub const IMPLEMENTS: Self = _

Symbol for the "implements" string.

pub const INTERFACE: Self = _

Symbol for the "interface" string.

pub const LET: Self = _

Symbol for the "let" string.

pub const PACKAGE: Self = _

Symbol for the "package" string.

pub const PRIVATE: Self = _

Symbol for the "private" string.

pub const PROTECTED: Self = _

Symbol for the "protected" string.

pub const PUBLIC: Self = _

Symbol for the "public" string.

pub const ANONYMOUS: Self = _

Symbol for the "anonymous" string.

pub const TRUE: Self = _

Symbol for the "true" string.

pub const FALSE: Self = _

Symbol for the "false" string.

pub const ASYNC: Self = _

Symbol for the "async" string.

pub const OF: Self = _

Symbol for the "of" string.

pub const TARGET: Self = _

Symbol for the "target" string.

pub const __PROTO__: Self = _

Symbol for the "__proto__" string.

pub(crate) fn new(value: usize) -> Option<Self>

Creates a new Sym from the provided value, or returns None if index is zero.

pub(crate) const unsafe fn new_unchecked(value: usize) -> Self

Creates a new Sym from the provided value, without checking if value is not zero

Safety

value must not be zero.

pub const fn get(self) -> usize

Returns the internal value of the Sym