Trait boa_interner::ToInternedString

pub trait ToInternedString {
    fn to_interned_string(&self, interner: &Interner) -> String;
}

Converts a given element to a string using an interner.

Required Methods

fn to_interned_string(&self, interner: &Interner) -> String

Converts a given element to a string using an interner.