Crate boa_interner

String interner for Boa.

The idea behind using a string interner is that in most of the code, strings such as identifiers and literals are often repeated. This causes extra burden when comparing them and storing them. A string interner stores a unique usize symbol for each string, making sure that there are no duplicates. This makes it much easier to compare, since it’s just comparing to usize, and also it’s easier to store, since instead of a heap-allocated string, you only need to store a usize. This reduces memory consumption and improves performance in the compiler.

Structs

Interner

The string interner for Boa.

Sym

The string symbol type for Boa.

Traits

ToInternedString

Converts a given element to a string using an interner.

Type Definitions

Backend

Backend of the string interner.