Struct boa_interner::Interner

pub struct Interner {
    pub(crate) utf8_interner: RawInterner<u8>,
    pub(crate) utf16_interner: RawInterner<u16>,
}

The string interner for Boa.

Fields
utf8_interner: RawInterner<u8>utf16_interner: RawInterner<u16>
Implementations

impl Interner

pub fn new() -> Self

Creates a new Interner.

pub fn with_capacity(capacity: usize) -> Self

Creates a new Interner with the specified capacity.

pub fn len(&self) -> usize

Returns the number of strings interned by the interner.

pub fn is_empty(&self) -> bool

Returns true if the Interner contains no interned strings.

pub fn get<'a, T>(&self, string: T) -> Option<Sym>where T: Into<JStrRef<'a>>,

Returns the symbol for the given string if any.

Can be used to query if a string has already been interned without interning.

pub fn get_or_intern<'a, T>(&mut self, string: T) -> Symwhere T: Into<JStrRef<'a>>,

Interns the given string.

Returns a symbol for resolution into the original string.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn get_or_intern_static( &mut self, utf8: &'static str, utf16: &'static [u16] ) -> Sym

Interns the given 'static string.

Returns a symbol for resolution into the original string.

Note

This is more efficient than Interner::get_or_intern, since it avoids allocating space for one string inside the Interner, with the disadvantage that you need to provide both the UTF-8 and the UTF-16 representation of the string.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn resolve(&self, symbol: Sym) -> Option<JSInternedStrRef<'_, '_>>

Returns the string for the given symbol if any.

Panics

Panics if the size of both statics is not equal or the interners do not have the same size

pub fn resolve_expect(&self, symbol: Sym) -> JSInternedStrRef<'_, '_>

Returns the string for the given symbol.

Panics

If the interner cannot resolve the given symbol.

pub(crate) fn get_common(string: JStrRef<'_>) -> Option<Sym>

Gets the symbol of the common string if one of them