Struct boa_interner::Interner

pub struct Interner {
    pub(crate) inner: StringInterner<BucketBackend<Sym>>,
}

The string interner for Boa.

This is a type alias that makes it easier to reference it in the code.

Fields
inner: StringInterner<BucketBackend<Sym>>
Implementations

impl Interner

pub fn with_capacity(cap: usize) -> Self

Creates a new StringInterner with the given initial capacity.

pub fn len(&self) -> usize

Returns the number of strings interned by the interner.

pub fn is_empty(&self) -> bool

Returns true if the string interner has no interned strings.

pub fn get<T>(&self, string: T) -> Option<Sym> where
T: AsRef<str>,

Returns the symbol for the given string if any.

Can be used to query if a string has already been interned without interning.

pub fn get_or_intern<T>(&mut self, string: T) -> Sym where
T: AsRef<str>,

Interns the given string.

Returns a symbol for resolution into the original string.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn get_or_intern_static(&mut self, string: &'static str) -> Sym

Interns the given 'static string.

Returns a symbol for resolution into the original string.

Note

This is more efficient than StringInterner::get_or_intern since it might avoid some memory allocations if the backends supports this.

Panics

If the interner already interns the maximum number of strings possible by the chosen symbol type.

pub fn shrink_to_fit(&mut self)

Shrink backend capacity to fit the interned strings exactly.

pub fn resolve(&self, symbol: Sym) -> Option<&str>

Returns the string for the given symbol if any.

pub fn resolve_expect(&self, symbol: Sym) -> &str

Returns the string for the given symbol.

Panics

If the interner cannot resolve the given symbol.

pub(crate) fn get_static(string: &str) -> Option<Sym>

Gets the symbol of the static string if one of them

impl Interner

pub(crate) const STATIC_STRINGS: [&'static str; 12]

List of commonly used static strings.

Make sure that any string added as a Sym constant is also added here.