Struct boa_interner::sym::Sym

pub struct Sym {
    value: NonZeroUsize,
}

The string symbol type for Boa.

This symbol type is internally a NonZeroUsize, which makes it pointer-width in size and it’s optimized so that it can occupy 1 pointer width even in an Option type.

Fields
value: NonZeroUsize
Implementations

impl Sym

pub(crate) fn new(value: usize) -> Option<Self>

Creates a new Sym from the provided value, or returns None if index is zero.

pub(crate) const unsafe fn new_unchecked(value: usize) -> Self

Creates a new Sym from the provided value, without checking if value is not zero

Safety

value must not be zero.

pub fn is_reserved_identifier(self) -> bool

Checks if this symbol is one of the reserved identifiers of the ECMAScript specification, excluding await and yield

pub fn is_strict_reserved_identifier(self) -> bool

Checks if this symbol is one of the strict reserved identifiers of the ECMAScript specification.

pub const fn get(self) -> usize

Returns the internal value of the Sym

impl Sym

pub const BREAK: Self = _

Symbol for the “break” string.

pub const CASE: Self = _

Symbol for the “case” string.

pub const CATCH: Self = _

Symbol for the “catch” string.

pub const CLASS: Self = _

Symbol for the “class” string.

pub const CONST: Self = _

Symbol for the “const” string.

pub const CONTINUE: Self = _

Symbol for the “continue” string.

pub const DEBUGGER: Self = _

Symbol for the “debugger” string.

pub const DEFAULT: Self = _

Symbol for the “default” string.

pub const DELETE: Self = _

Symbol for the “delete” string.

pub const DO: Self = _

Symbol for the “do” string.

pub const ELSE: Self = _

Symbol for the “else” string.

pub const ENUM: Self = _

Symbol for the “enum” string.

pub const EXPORT: Self = _

Symbol for the “export” string.

pub const EXTENDS: Self = _

Symbol for the “extends” string.

pub const FALSE: Self = _

Symbol for the “false” string.

pub const FINALLY: Self = _

Symbol for the “finally” string.

pub const FOR: Self = _

Symbol for the “for” string.

pub const FUNCTION: Self = _

Symbol for the “function” string.

pub const IF: Self = _

Symbol for the “if” string.

pub const IMPORT: Self = _

Symbol for the “import” string.

pub const IN: Self = _

Symbol for the “in” string.

pub const INSTANCEOF: Self = _

Symbol for the “instanceof” string.

pub const NEW: Self = _

Symbol for the “new” string.

pub const NULL: Self = _

Symbol for the “null” string.

pub const RETURN: Self = _

Symbol for the “return” string.

pub const SUPER: Self = _

Symbol for the “super” string.

pub const SWITCH: Self = _

Symbol for the “switch” string.

pub const THIS: Self = _

Symbol for the “this” string.

pub const THROW: Self = _

Symbol for the “throw” string.

pub const TRUE: Self = _

Symbol for the “true” string.

pub const TRY: Self = _

Symbol for the “try” string.

pub const TYPEOF: Self = _

Symbol for the “typeof” string.

pub const VAR: Self = _

Symbol for the “var” string.

pub const VOID: Self = _

Symbol for the “void” string.

pub const WHILE: Self = _

Symbol for the “while” string.

pub const WITH: Self = _

Symbol for the “with” string.

pub const IMPLEMENTS: Self = _

Symbol for the “implements” string.

pub const INTERFACE: Self = _

Symbol for the “interface” string.

pub const LET: Self = _

Symbol for the “let” string.

pub const PACKAGE: Self = _

Symbol for the “package” string.

pub const PRIVATE: Self = _

Symbol for the “private” string.

pub const PROTECTED: Self = _

Symbol for the “protected” string.

pub const PUBLIC: Self = _

Symbol for the “public” string.

pub const STATIC: Self = _

Symbol for the “static” string.

pub const YIELD: Self = _

Symbol for the “yield” string.

pub const EMPTY_STRING: Self = _

Symbol for the “” string.

pub const PROTOTYPE: Self = _

Symbol for the “prototype” string.

pub const CONSTRUCTOR: Self = _

Symbol for the “constructor” string.

pub const ARGUMENTS: Self = _

Symbol for the “arguments” string.

pub const EVAL: Self = _

Symbol for the “eval” string.

pub const REGEXP: Self = _

Symbol for the “RegExp” string.

pub const GET: Self = _

Symbol for the “get” string.

pub const SET: Self = _

Symbol for the “set” string.

pub const MAIN: Self = _

Symbol for the “<main>” string.

pub const RAW: Self = _

Symbol for the “raw” string.

pub const ANONYMOUS: Self = _

Symbol for the “anonymous” string.

pub const ASYNC: Self = _

Symbol for the “async” string.

pub const OF: Self = _

Symbol for the “of” string.

pub const TARGET: Self = _

Symbol for the “target” string.

pub const AS: Self = _

Symbol for the “as” string.

pub const FROM: Self = _

Symbol for the “from” string.

pub const __PROTO__: Self = _

Symbol for the “proto” string.

pub const NAME: Self = _

Symbol for the “name” string.

pub const AWAIT: Self = _

Symbol for the “await” string.

pub const DEFAULT_EXPORT: Self = _

Symbol for the “*default*” string.

pub const META: Self = _

Symbol for the “meta” string.